third_party/WebKit/LayoutTests/editing/selection/select_all/select_all_contenteditable.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="div" contenteditable="true"></div>
<script>
test(function() {
    var div = document.getElementById('div');
    div.focus();
    assert_false(document.queryCommandEnabled('SelectAll'));

    div.appendChild(document.createTextNode('x'));
    div.focus();
    assert_true(document.queryCommandEnabled('SelectAll'));
}, 'SelectAll is not enabled if contenteditable is empty');
</script>
