third_party/WebKit/LayoutTests/platform/win/editing/selection/select_all/select_all_contenteditable-expected.txt

This is a testharness.js-based test.
PASS SelectAll is always enabled for contenteditable 
Harness: the test ran to completion.
